services/mock-consumer/requirements.txt
fluent-logger==0.11.1
stomp.py==8.1.2